# backend/app/APIspecification.yaml
components:
  schemas:
    Body_analyze_student_report_v1_analyze_student_report__post:
      properties:
        file:
          format: binary
          title: File
          type: string
        user_id:
          title: User Id
          type: integer
      required:
      - user_id
      - file
      title: Body_analyze_student_report_v1_analyze_student_report__post
      type: object
    Body_generate_milestone_v1_generate_milestone__post:
      properties:
        file:
          format: binary
          title: File
          type: string
      required:
      - file
      title: Body_generate_milestone_v1_generate_milestone__post
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: A web-based application designed to assist instructors in tracking
    student projects using AI.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AI-Assisted Project Tracking System
  version: 1.0.0
openapi: 3.1.0
paths:
  /:
    get:
      operationId: read_root__get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Read Root
  /v1/analyze-code/:
    post:
      operationId: analyze_code_v1_analyze_code__post
      parameters:
      - in: query
        name: username
        required: true
        schema:
          title: Username
          type: string
      - in: query
        name: repo_name
        required: true
        schema:
          title: Repo Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Analyze Code
      tags:
      - Analyze codes from github
  /v1/analyze-student-report/:
    post:
      operationId: analyze_student_report_v1_analyze_student_report__post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_analyze_student_report_v1_analyze_student_report__post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Analyze Student Report
      tags:
      - Analyze pdf documents
  /v1/best-practice/:
    get:
      operationId: best_practices_v1_best_practice__get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Best Practices
      tags:
      - Generate best practices for a given language
  /v1/generate_milestone/:
    post:
      operationId: generate_milestone_v1_generate_milestone__post
      parameters:
      - in: query
        name: project_title
        required: true
        schema:
          title: Project Title
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_generate_milestone_v1_generate_milestone__post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Generate Milestone
      tags:
      - generate milestones
  /v1/get-all-branches/:
    get:
      operationId: get_all_branches_v1_get_all_branches__get
      parameters:
      - in: query
        name: username
        required: true
        schema:
          title: Username
          type: string
      - in: query
        name: repo_name
        required: true
        schema:
          title: Repo Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get All Branches
      tags:
      - Analyze codes from github
  /v1/get-all-repo/:
    get:
      operationId: get_all_repo_v1_get_all_repo__get
      parameters:
      - in: query
        name: username
        required: true
        schema:
          title: Username
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get All Repo
      tags:
      - Analyze codes from github